+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 ## Coming in [v3.0.2](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v3.0.1)
+### Other
+- Import: When an image file already exists it is now replaced and the import of the mod is not canceled
+
 ### Bug fixes
 - "Import selected" button now works properly and the action will no longer fail.
 - One translation key was not found